Refactor maximum/minimum
