+++ source/framework/python/src/algorithm/local_operation.cpp
@@ -48,8 +48,6 @@ namespace lue::framework {
     void bind_logical_exclusive_or(pybind11::module& module);
     void bind_logical_inclusive_or(pybind11::module& module);
     void bind_logical_not(pybind11::module& module);
-    void bind_maximum(pybind11::module& module);
-    void bind_minimum(pybind11::module& module);
     void bind_multiply(pybind11::module& module);
     void bind_normal(pybind11::module& module);
     void bind_not_equal_to(pybind11::module& module);
@@ -112,8 +110,6 @@ namespace lue::framework {
         bind_logical_exclusive_or(module);
         bind_logical_inclusive_or(module);
         bind_logical_not(module);
-        bind_maximum(module);
-        bind_minimum(module);
         bind_multiply(module);
         bind_normal(module);
         bind_not_equal_to(module);
